вынос в молули shuffle и getComments
