+++ js/shuffle.js
@@ -1,0 +1,7 @@
+// Тасование Фишера — Йетса: https://learn.javascript.ru/task/shuffle
+export const shuffle = (array) => {
+  for (let i = array.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [array[i], array[j]] = [array[j], array[i]];
+  }
+};
